Implement requests to split files
